Moving the disjunction function into core functions
and minor corrections
